requirements.txt
nltk
newspaper3k
requests
bs4
datetime
pandas